+++ src/pages/creer-une-histoire/1.tsx
@@ -3,10 +3,10 @@ import { useRouter } from 'next/router';
 import React from 'react';
 
 import AddIcon from '@mui/icons-material/Add';
-import { TextField, Grid, ButtonBase } from '@mui/material';
+import { Grid, ButtonBase, FormControl, FormHelperText, InputLabel, MenuItem, Select } from '@mui/material';
 
 import { isStory } from 'src/activity-types/anyActivity';
-import { DEFAULT_STORY_DATA } from 'src/activity-types/story.constants';
+import { DEFAULT_STORY_DATA, ODD_CHOICE } from 'src/activity-types/story.constants';
 import { Base } from 'src/components/Base';
 import { KeepRatio } from 'src/components/KeepRatio';
 import { Steps } from 'src/components/Steps';
@@ -17,18 +17,19 @@ import { DeleteButton } from 'src/components/buttons/DeleteButton';
 import { ActivityContext } from 'src/contexts/activityContext';
 import { VillageContext } from 'src/contexts/villageContext';
 import { useImageStoryRequests } from 'src/services/useImagesStory';
-import { bgPage, primaryColor, errorColor } from 'src/styles/variables.const';
+import { primaryColor, bgPage, errorColor } from 'src/styles/variables.const';
 import { ActivityStatus, ActivityType } from 'types/activity.type';
-import type { StoriesData, StoryElement } from 'types/story.type';
+import type { StoriesData } from 'types/story.type';
 
 const StoryStep1 = () => {
   const router = useRouter();
-  const { activity, updateActivity, createNewActivity, save } = React.useContext(ActivityContext);
+  const { activity, createNewActivity, updateActivity, save } = React.useContext(ActivityContext);
   const { selectedPhase } = React.useContext(VillageContext);
   const { deleteStoryImage } = useImageStoryRequests();
   const [isError, setIsError] = React.useState<boolean>(false);
   const [isImageModalOpen, setIsImageModalOpen] = React.useState(false);
-  const data = (activity?.data as StoriesData) || null;
+  const [oDDChoice, setODDChoice] = React.useState('');
+  const data = (activity?.data as StoriesData) || { odd: ODD_CHOICE[0] };
   const isEdit = activity !== null && activity.status !== ActivityStatus.DRAFT;
 
   // Create the story activity.
@@ -52,18 +53,10 @@ const StoryStep1 = () => {
     }
   }, [activity, createNewActivity, router.query, selectedPhase]);
 
-  const dataChange = (key: keyof StoryElement) => (event: React.ChangeEvent<HTMLInputElement>) => {
-    const value = event.target.value.slice(0, 400);
-    const { object } = data;
-    const newData = { ...data, object: { ...object, [key]: value } };
-    updateActivity({ data: newData });
-  };
-
-  // Update the "object step" image url, when upload an image.
+  // Update the "odd step" image url, when upload an image.
   const setImage = (imageUrl: string) => {
-    const { object } = data;
-    // imageId = 0 when we are changing the image of the object step.
-    updateActivity({ data: { ...data, object: { ...object, imageUrl, imageId: 0 } } });
+    const { odd } = data;
+    updateActivity({ data: { ...data, odd: { ...odd, inspiredStoryId: activity?.id, imageUrl, imageId: 0 } } });
   };
 
   React.useEffect(() => {
@@ -86,16 +79,16 @@ const StoryStep1 = () => {
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
         {!isEdit && <BackButton href="/creer-une-histoire" />}
         <Steps
-          steps={['Objet', 'Lieu', 'ODD', 'Histoire', 'Prévisualisation']}
+          steps={['ODD', 'Objet', 'Lieu', 'Histoire', 'Prévisualisation']}
           urls={['/creer-une-histoire/1?edit', '/creer-une-histoire/2', '/creer-une-histoire/3', '/creer-une-histoire/4', '/creer-une-histoire/5']}
           activeStep={0}
         />
         <div className="width-900">
-          <h1>Inventez et dessinez un objet magique</h1>
+          <h1>Choisissez et dessinez l’objectif du développement durable atteint</h1>
           <p className="text">
-            Cet objet, tout comme le lieu que vous choisirez à l’étape suivante, est magique ! Grâce à leurs pouvoirs, le village idéal a atteint
-            l’objectif du développement durable que vous choisirez en étape 3.
+            Grâce aux pouvoirs magiques de l’objet et du lieu choisis aux étapes précédentes, un des objectifs du développement durable a été atteint.
           </p>
+          <p className="text">Choisissez lequel et dessinez-le.</p>
           <Grid container spacing={3}>
             <Grid item xs={12} md={6}>
               <div style={{ marginTop: '1.5rem' }}>
@@ -113,19 +106,25 @@ const StoryStep1 = () => {
                           justifyContent: 'center',
                         }}
                       >
-                        {data?.object?.imageUrl ? (
-                          <Image layout="fill" objectFit="cover" alt="image de l'objet" src={data?.object?.imageUrl} unoptimized />
+                        {data.odd?.imageUrl ? (
+                          <Image
+                            layout="fill"
+                            objectFit="cover"
+                            alt="image de l'objectif de développement durable"
+                            src={data.odd?.imageUrl}
+                            unoptimized
+                          />
                         ) : (
                           <AddIcon style={{ fontSize: '80px' }} />
                         )}
                       </div>
                     </KeepRatio>
                   </ButtonBase>
-                  {data?.object?.imageUrl && (
+                  {data.odd?.imageUrl && (
                     <div style={{ position: 'absolute', top: '0.25rem', right: '0.25rem' }}>
                       <DeleteButton
                         onDelete={() => {
-                          deleteStoryImage(data.object.imageId, data, 1);
+                          deleteStoryImage(data.odd.imageId, data, 3);
                           setImage('');
                         }}
                         confirmLabel="Êtes-vous sur de vouloir supprimer l'image ?"
@@ -138,41 +137,40 @@ const StoryStep1 = () => {
                     id={0}
                     isModalOpen={isImageModalOpen}
                     setIsModalOpen={setIsImageModalOpen}
-                    imageUrl={data?.object?.imageUrl || ''}
+                    imageUrl={data.odd?.imageUrl || ''}
                     setImageUrl={setImage}
                   />
                 </div>
+                <FormControl variant="outlined" className="full-width" style={{ marginTop: '1rem' }}>
+                  <InputLabel id="select-ODD">ODD</InputLabel>
+                  <Select
+                    error={isError && data?.odd?.description === ''}
+                    labelId="select-ODD"
+                    id="select-ODD-outlined"
+                    value={oDDChoice || data.odd?.description}
+                    onChange={(event) => {
+                      setODDChoice(event.target.value as string);
+                      const { odd } = data;
+                      updateActivity({ data: { ...data, odd: { ...odd, description: event.target.value } } });
+                    }}
+                    label="Village"
+                  >
+                    {(ODD_CHOICE || []).map((v, index) => (
+                      <MenuItem value={v.choice} key={index + 1}>
+                        {v.choice}
+                      </MenuItem>
+                    ))}
+                  </Select>
+                  <FormHelperText>Choisissez votre ODD </FormHelperText>
+                </FormControl>
               </div>
-              <TextField
-                error={isError && data?.object?.description === ''}
-                helperText={isError && data?.object?.description === '' && 'Écrivez la description de votre image !'}
-                id="standard-multiline-static"
-                label="Décrivez l’objet magique"
-                value={data?.object?.description || ''}
-                onChange={dataChange('description')}
-                variant="outlined"
-                multiline
-                maxRows={4}
-                style={{ width: '100%', marginTop: '25px', color: 'primary' }}
-                inputProps={{
-                  maxLength: 400,
-                }}
-              />
-              {data?.object?.description ? (
-                <div style={{ width: '100%', textAlign: 'right' }}>
-                  <span className="text text--small">{data.object.description.length}/400</span>
-                </div>
-              ) : (
-                <div style={{ width: '100%', textAlign: 'right' }}>
-                  <span className="text text--small">0/400</span>
-                </div>
-              )}
             </Grid>
           </Grid>
         </div>
-        <StepsButton next={onNext} />
+        <StepsButton prev="/creer-une-histoire/2" next={onNext} />
       </div>
     </Base>
   );
 };
+
 export default StoryStep1;
