+++ src/pages/creer-une-histoire/2.tsx
@@ -15,7 +15,7 @@ import { getErrorSteps } from 'src/components/activities/storyChecks';
 import { DeleteButton } from 'src/components/buttons/DeleteButton';
 import { ActivityContext } from 'src/contexts/activityContext';
 import { useImageStoryRequests } from 'src/services/useImagesStory';
-import { primaryColor, bgPage } from 'src/styles/variables.const';
+import { bgPage, primaryColor } from 'src/styles/variables.const';
 import type { StoriesData, StoryElement } from 'types/story.type';
 
 const StoryStep2 = () => {
@@ -27,7 +27,7 @@ const StoryStep2 = () => {
 
   const errorSteps = React.useMemo(() => {
     if (data !== null) {
-      return getErrorSteps(data.object, 1);
+      return getErrorSteps(data.odd, 1);
     }
     return [];
   }, [data]);
@@ -42,23 +42,16 @@ const StoryStep2 = () => {
 
   const dataChange = (key: keyof StoryElement) => (event: React.ChangeEvent<HTMLInputElement>) => {
     const value = event.target.value.slice(0, 400);
-    const { place } = data;
-    const newData = { ...data, place: { ...place, [key]: value } };
+    const { object } = data;
+    const newData = { ...data, object: { ...object, [key]: value } };
     updateActivity({ data: newData });
   };
 
   // Update the "object step" image url, when upload an image.
   const setImage = (imageUrl: string) => {
-    const { object, place } = data;
+    const { object } = data;
     // imageId = 0 when we are changing the image of the object step.
-    updateActivity({
-      data: {
-        ...data,
-        object: { ...object, inspiredStoryId: activity?.id },
-        place: { ...place, inspiredStoryId: activity?.id, imageUrl, imageId: 0 },
-        isOriginal: true,
-      },
-    });
+    updateActivity({ data: { ...data, object: { ...object, imageUrl, imageId: 0 } } });
   };
 
   const onNext = () => {
@@ -69,21 +62,20 @@ const StoryStep2 = () => {
   if (data === null || activity === null || !isStory(activity)) {
     return <div></div>;
   }
-
   return (
     <Base>
       <div style={{ width: '100%', padding: '0.5rem 1rem 1rem 1rem' }}>
         <Steps
-          steps={['Objet', 'Lieu', 'ODD', 'Histoire', 'Prévisualisation']}
+          steps={['ODD', 'Objet', 'Lieu', 'Histoire', 'Prévisualisation']}
           urls={['/creer-une-histoire/1?edit', '/creer-une-histoire/2', '/creer-une-histoire/3', '/creer-une-histoire/4', '/creer-une-histoire/5']}
           activeStep={1}
           errorSteps={errorSteps}
         />
         <div className="width-900">
-          <h1>Inventez et dessinez un lieu extraordinaire</h1>
+          <h1>Inventez et dessinez un objet magique</h1>
           <p className="text">
-            Ce lieu, tout comme l’objet que vous avez choisi à l’étape précédente, est extraodinaire ! Grâce à leurs pouvoirs, le village idéal a
-            atteint l’objectif du développement durable que vous choisirez en étape 3.
+            Cet objet, tout comme le lieu que vous choisirez à l’étape suivante, est magique ! Grâce à leurs pouvoirs, le village idéal a atteint
+            l’objectif du développement durable que vous choisirez en étape 3.
           </p>
           <Grid container spacing={3}>
             <Grid item xs={12} md={6}>
@@ -102,19 +94,19 @@ const StoryStep2 = () => {
                           justifyContent: 'center',
                         }}
                       >
-                        {data.place?.imageUrl ? (
-                          <Image layout="fill" objectFit="cover" alt="image du lieu" src={data.place?.imageUrl} unoptimized />
+                        {data?.object?.imageUrl ? (
+                          <Image layout="fill" objectFit="cover" alt="image de l'objet" src={data?.object?.imageUrl} unoptimized />
                         ) : (
                           <AddIcon style={{ fontSize: '80px' }} />
                         )}
                       </div>
                     </KeepRatio>
                   </ButtonBase>
-                  {data.place?.imageUrl && (
+                  {data?.object?.imageUrl && (
                     <div style={{ position: 'absolute', top: '0.25rem', right: '0.25rem' }}>
                       <DeleteButton
                         onDelete={() => {
-                          deleteStoryImage(data.place.imageId, data, 2);
+                          deleteStoryImage(data.object.imageId, data, 1);
                           setImage('');
                         }}
                         confirmLabel="Êtes-vous sur de vouloir supprimer l'image ?"
@@ -127,15 +119,16 @@ const StoryStep2 = () => {
                     id={0}
                     isModalOpen={isImageModalOpen}
                     setIsModalOpen={setIsImageModalOpen}
-                    imageUrl={data.place?.imageUrl || ''}
+                    imageUrl={data?.object?.imageUrl || ''}
                     setImageUrl={setImage}
                   />
                 </div>
               </div>
               <TextField
+                helperText={'Écrivez la description de votre image !'}
                 id="standard-multiline-static"
-                label="Décrivez le lieu extraordinaire"
-                value={data.place?.description || ''}
+                label="Décrivez l’objet magique"
+                value={data?.object?.description || ''}
                 onChange={dataChange('description')}
                 variant="outlined"
                 multiline
@@ -145,9 +138,9 @@ const StoryStep2 = () => {
                   maxLength: 400,
                 }}
               />
-              {data.place?.description ? (
+              {data?.object?.description ? (
                 <div style={{ width: '100%', textAlign: 'right' }}>
-                  <span className="text text--small">{data.place.description.length}/400</span>
+                  <span className="text text--small">{data.object.description.length}/400</span>
                 </div>
               ) : (
                 <div style={{ width: '100%', textAlign: 'right' }}>
@@ -157,10 +150,9 @@ const StoryStep2 = () => {
             </Grid>
           </Grid>
         </div>
-        <StepsButton prev={`/creer-une-histoire/1?edit=${activity.id}`} next={onNext} />
+        <StepsButton next={onNext} />
       </div>
     </Base>
   );
 };
-
 export default StoryStep2;
